refactor(clicker): dedupe glob-image resolver and extract useRotatingCursors
Collapse the four copies of the eager import.meta.glob + filename lookup
into a single getGlobImage helper, and move the RotatingCursors Web
Animations engine out of the SFC into a useRotatingCursors composable.
