+++ packages/app/app/services/clicker/getGlobImage.ts
@@ -1,0 +1,4 @@
+import { filename } from "pathe/utils";
+
+export const getGlobImage = (glob: Record<string, string>, id: string) =>
+  Object.fromEntries(Object.entries(glob).map(([path, value]) => [filename(path), value]))[id];
